Refactor Certificate class and add ChainItem class
In the Certificate class, we expanded its functionality to implement Arrayable interface and added data validation in the constructor. Additionally, we introduced a new ChainItem class which holds information about individual parts of the certificate chain. This aims to provide a more granular and controlled access over individual components of the SSL certificate chain. The toArray methods in both classes were also annotated with #[\Override] attribute.