Refactor certificate tests with more detailed data
The certificate tests have been refactored to use a more detailed data configuration. With the more thorough setup, the tests now appropriately check the functionality of each certificate property. A comment has been left in place for a test that still needs fixing, aimed at getting the certificate data as an array.