Refactor Config class and update ConfigInterface
The Config class has been refactored to define listeners with an initial empty array, enhance the 'toArray' method and correct the 'toJson' method. A 'toJson' method contract has also been added to the ConfigInterface. Also, added a 'getName' method to the Listener class. This revision improves the efficiency and correctness of the configuration handling.