Optimize import statements and update docblocks in PHP code
This commit focuses on source code optimization and improving readability in the PHP files. In "Certificate.php" and "JsModule.php", specific classes are added to the import statements at the top of the file, which simplifies their usage in the code. Docblocks in "Unit.php" and "ConfigInterface.php" are updated to make them more precise and less redundant.