Implement JSON serialization in interfaces and classes
Updated several classes and interfaces to extend the Arrayable and Jsonable interfaces. This allows for consistent serialization of objects to JSON format. Functionality was added to the Certificate.php, StatisticsInterface.php, UnitInterface.php, Statistics.php, and Unit.php files. Furthermore, the 'chain' property in Certificate.php has been mapped to an array for easy conversion to JSON.