Remove files associated with Nginx Unit PHP SDK documentation
This commit removes all files related to the documentation of the Nginx Unit PHP SDK. This includes files detailing installation, basic usage, authors, changelog, and more. This action corresponds to the decision to move the documentation elsewhere or to discontinue maintaining it within this project's repository. It helps to reduce clutter and streamlines the repository structure.